Split prompt template helpers
