+++ tach.toml
@@ -70,7 +70,7 @@ depends_on = [
     "mindroom.logging_config",
     "mindroom.matrix.identity",
     "mindroom.model_loading",
-    "mindroom.prompts",
+    "mindroom.prompt_templates",
     "mindroom.runtime_resolution",
     "mindroom.timing",
     "mindroom.tool_system.catalog",
@@ -141,7 +141,6 @@ depends_on = [
     "mindroom.logging_config",
     "mindroom.media_fallback",
     "mindroom.media_inputs",
-    "mindroom.prompts",
 ]
 visibility = [
     "mindroom.ai",
@@ -640,7 +639,6 @@ depends_on = [
     "mindroom.constants",
     "mindroom.history",
     "mindroom.matrix.client_visible_messages",
-    "mindroom.prompts",
     "mindroom.streaming",
 ]
 
@@ -664,7 +662,6 @@ depends_on = [
     "mindroom.hooks",
     "mindroom.logging_config",
     "mindroom.metadata_merge",
-    "mindroom.prompts",
     "mindroom.timing",
     "mindroom.token_budget",
     "mindroom.tool_system.runtime_context",
@@ -969,6 +966,7 @@ depends_on = [
     "mindroom.logging_config",
     "mindroom.matrix_identifiers",
     "mindroom.mcp.config",
+    "mindroom.prompt_templates",
     "mindroom.prompts",
     "mindroom.tool_system.catalog",
     "mindroom.tool_system.plugin_imports",
@@ -1071,22 +1069,26 @@ visibility = ["mindroom.memory.functions"]
 
 [[modules]]
 path = "mindroom.memory._prompting"
-depends_on = ["mindroom.memory._shared", "mindroom.prompts"]
+depends_on = ["mindroom.memory._shared", "mindroom.prompt_templates"]
 visibility = ["mindroom.memory", "mindroom.memory.functions"]
+
+[[modules]]
+path = "mindroom.prompt_templates"
+depends_on = ["mindroom.prompts"]
+visibility = [
+    "mindroom.agents",
+    "mindroom.config.main",
+    "mindroom.memory._prompting",
+]
 
 [[modules]]
 path = "mindroom.prompts"
 depends_on = []
 visibility = [
-    "mindroom.agents",
-    "mindroom.ai_runtime",
     "mindroom.avatar_generation",
     "mindroom.codex_model",
     "mindroom.config.main",
-    "mindroom.execution_preparation",
-    "mindroom.history.compaction",
-    "mindroom.media_fallback",
-    "mindroom.memory._prompting",
+    "mindroom.prompt_templates",
 ]
 
 [[modules]]
